Conditional rendering for instructor dashboard
